draw a line between sys/ and asm/; make use of generic arch
restructure the arch/* directory to differentiate the functionality
of sys/ and asm/. In which the former refer to arch-specific kernel
service; the latter refer to code related to low-level architectural
feature.

make use the arch/generic which will serve as a supplier of default
implementation, to allow sharing common code across different arch.